+++ src/os2datascanner/projects/report/reportapp/models/documentreport.py
@@ -58,7 +58,9 @@ class DocumentReport(models.Model):
 
     # the name of the specific resource a handle points to. The equivalent of a #
     # filename
-    name = models.CharField(max_length=256, verbose_name=_("name"), default="")
+    name = models.CharField(max_length=256, verbose_name=_("name"), default="", db_index=True)
+
+    place = models.CharField(max_length=256, verbose_name=_("place"), default="", db_index=True)
 
     source_type = models.CharField(max_length=2000,
                                    verbose_name=_("source type"), db_index=True)